changed confidence_level to float field)

+++ assassin_server/assassin/models.py
@@ -16,7 +16,7 @@ class Attempt(models.Model):
     image = models.FileField(upload_to='attempts')
     comment = models.CharField(max_length=200, blank=True, null=True)
     created_date = models.DateTimeField(auto_now_add=True)
-    confidence_level = models.IntegerField(blank=True, null=True)
+    confidence_level = models.FloatField(blank=True, null=True)
 
     def save(self):
         # Call save here so that the file has a path that cv2 can read from
